Add start screen as home, move match screen to /match
The app now opens on a new start screen after onboarding:
- Time-of-day dependent greeting (variants rotate daily)
- "Neue Partie starten" for the last played game
- Button to the active (or most recently played) match
- Schnelles Spiel and Meine Spiele shortcuts
- Freunde/Timer/Würfel shortcuts plus a small stats strip

The sidebar shows the new Start tab instead of Aktuelle Partie - the
running match is reached from the start screen or a game's Partien
list (route moved from / to /match).

The match setup phase now shows a single "Spieler hinzufügen" button
while no players are seated (the second one appears with the list).
