+++ apps/score-tracker/maestro-tests/src/tests/setup-flow-test.ts
@@ -20,6 +20,12 @@ test
 	// First launch shows the onboarding tour - skip it (no-op when already dismissed)
 	.optionalTapOnId(ComponentIds.ONBOARDING_SKIP_BUTTON)
 	.waitForAnimationToEnd()
+
+	// The app opens on the start screen - "Schnelles Spiel" jumps into the
+	// setup phase of a fresh match
+	.takeScreenshot('start-screen')
+	.tapOnId(ComponentIds.START_SCREEN_QUICK_MATCH_ROW)
+	.waitForAnimationToEnd()
 	.takeScreenshot('setup-empty')
 
 	// Add two guest players - the chooser stays open between adds and shows
